# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Signal87 Core API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /documents/{id}/chat:
    post:
      operationId: chatWithDocument
      tags: [chat]
      summary: Ask a question about a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatInput"
      responses:
        "200":
          description: Answer with citations and debug info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"
        "404":
          description: Document not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Document has no readable text to answer from
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /documents/brief:
    post:
      operationId: generateBrief
      tags: [chat]
      summary: Generate a structured executive brief over 1-5 selected documents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BriefInput"
      responses:
        "200":
          description: Structured brief with citations and verification trace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BriefResult"
        "400":
          description: Invalid selection or brief type (e.g. Comparison Brief with fewer than 2 documents, or a selected document has no chunks)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: One or more selected documents not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /documents/{id}/history:
    get:
      operationId: getChatHistory
      tags: [chat]
      summary: Get chat history for a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Chat history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
    delete:
      operationId: clearChatHistory
      tags: [chat]
      summary: Clear chat history for a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: History cleared
...
components:
  schemas:
...
    ChatInput:
      type: object
      required: [question]
      properties:
        question:
          type: string
          minLength: 1
    Citation:
      type: object
      required: [chunkIndex, content, relevanceScore]
      properties:
        chunkIndex:
          type: integer
        content:
          type: string
        relevanceScore:
          type: number
    DebugInfo:
      type: object
      required:
        - route
        - provider
        - model
        - fallbackUsed
        - documentId
        - chunksSearched
        - chunksRetrieved
        - retrievalLatencyMs
        - llmLatencyMs
        - totalLatencyMs
      properties:
        route:
          type: string
        provider:
          type: string
        model:
          type: string
        fallbackUsed:
          type: boolean
        documentId:
          type: integer
        chunksSearched:
          type: integer
        chunksRetrieved:
          type: integer
        retrievalLatencyMs:
          type: number
        llmLatencyMs:
          type: number
        totalLatencyMs:
          type: number
        errors:
          type: ["string", "null"]
    ChatResponse:
      type: object
      required: [answer, citations, debug]
      properties:
        answer:
          type: string
        citations:
          type: array
          items:
            $ref: "#/components/schemas/Citation"
        debug:
          $ref: "#/components/schemas/DebugInfo"
...
    MultiChatDocumentStat:
      type: object
      required: [documentId, documentName, chunksSearched, chunksRetrieved]
      properties:
        documentId:
          type: integer
        documentName:
          type: string
        chunksSearched:
          type: integer
        chunksRetrieved:
          type: integer
...
    BriefInput:
      type: object
      required: [documentIds, briefType]
      properties:
        documentIds:
          type: array
          minItems: 1
          maxItems: 5
          items:
            type: integer
        briefType:
          type: string
          enum: [executive_summary, risk, diligence, contract_review, comparison]
        focus:
          type: ["string", "null"]
          maxLength: 500
    BriefSection:
      type: object
      required: [heading, body]
      properties:
        heading:
          type: string
        body:
          type: string
    BriefCitation:
      type: object
      required: [citationNumber, documentId, documentName, chunkIndex, content, relevanceScore]
      properties:
        citationNumber:
          type: integer
        documentId:
          type: integer
        documentName:
          type: string
        chunkIndex:
          type: integer
        content:
          type: string
        relevanceScore:
          type: number
    BriefDebugInfo:
      type: object
      required:
        - route
        - provider
        - model
        - fallbackUsed
        - briefType
        - focusProvided
        - documentIds
        - documentNames
        - documentsSearched
        - chunksSearched
        - chunksRetrieved
        - chunksRetrievedByDocument
        - retrievalLatencyMs
        - llmLatencyMs
        - totalLatencyMs
      properties:
        route:
          type: string
        provider:
          type: string
        model:
          type: string
        fallbackUsed:
          type: boolean
        briefType:
          type: string
        focusProvided:
          type: boolean
        documentIds:
          type: array
          items:
            type: integer
        documentNames:
          type: array
          items:
            type: string
        documentsSearched:
          type: integer
        chunksSearched:
          type: integer
        chunksRetrieved:
          type: integer
        chunksRetrievedByDocument:
          type: array
          items:
            $ref: "#/components/schemas/MultiChatDocumentStat"
        retrievalLatencyMs:
          type: number
        llmLatencyMs:
          type: number
        totalLatencyMs:
          type: number
        errors:
          type: ["string", "null"]
    BriefResult:
      type: object
      required: [briefType, title, sections, citations, debug]
      properties:
        briefType:
          type: string
        title:
          type: string
        sections:
          type: array
          items:
            $ref: "#/components/schemas/BriefSection"
        citations:
          type: array
          items:
            $ref: "#/components/schemas/BriefCitation"
        debug:
          $ref: "#/components/schemas/BriefDebugInfo"
    ChatMessage:
      type: object
      required: [id, documentId, role, content, createdAt]
      properties:
        id:
          type: integer
        documentId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
        debug:
          type: ["string", "null"]
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string